src/extractor/requirements.txt
websocket-client
dotenv
confluent-kafka